+++ src/runtime_src/driver/xclng/xrt/user_gem/xbsak_debug.cpp
@@ -126,7 +126,7 @@ int xcldev::device::readSPMCounters() {
             << "  " << std::setw(16)  << "Last Rd Addr"
             << "  " << std::setw(16)  << "Last Rd Data"
             << std::endl;
-    for (int i = 0; i<debugResults.NumSlots; ++i) {
+    for (size_t i = 0; i<debugResults.NumSlots; ++i) {
         std::cout << std::left
             << std::setw(col1) << cuNameportNames[i].first
             << " " << std::setw(col2) << cuNameportNames[i].second
@@ -165,7 +165,7 @@ int xcldev::device::readLAPCheckers(int aVerbose) {
     int col1 = std::max(widths.first, strlen("CU Name")) + 4;
     int col2 = std::max(widths.second, strlen("AXI Portname"));
 
-    for (int i = 0; i<debugResults.NumSlots; ++i) {
+    for (size_t i = 0; i<debugResults.NumSlots; ++i) {
         if (!xclAXICheckerCodes::isValidAXICheckerCodes(debugResults.OverallStatus[i],
                         debugResults.SnapshotStatus[i], debugResults.CumulativeStatus[i])) {
             std::cout << "CU Name: " << cuNameportNames[i].first << " AXI Port: " << cuNameportNames[i].second << "\n";
@@ -208,7 +208,7 @@ int xcldev::device::readLAPCheckers(int aVerbose) {
                 << "  " << std::setw(16) << "Cumulative[2]"
                 << "  " << std::setw(16) << "Cumulative[3]"
                 << std::endl;
-        for (int i = 0; i<debugResults.NumSlots; ++i) {
+        for (size_t i = 0; i<debugResults.NumSlots; ++i) {
             std::cout << std::left
                 << std::setw(col1) << cuNameportNames[i].first
                 << " " << std::setw(col2) << cuNameportNames[i].second
